Wheeler 1: extract src/js into core/ + crypto/ (prefactor)
Mechanical restructure ahead of the Wheeler (TradFi) sibling app: move the
platform-neutral modules under src/js/core/ and the crypto-specific ones
(wallet popup, cloud-sync, chain-sync, chain-apply) under src/js/crypto/.

build.py and test/helpers/loadApp.js now glob both dirs and concatenate by
basename, so the numeric prefixes still drive order regardless of directory.
Test require() paths updated to the new locations.

HyperWheel artifact is byte-identical (same sha) and npm test is fully green.

Closes #87
