+++ src/js/core/04-trade-crud.js
