+++ src/js/core/05a-merge-open-lots.js
