+++ src/js/core/05c-outcome-distribution.js
