+++ src/js/core/05d-calc-stats.js
