+++ src/js/core/06-render-table.js
