+++ src/js/core/06a-render-outcome-chart.js
