+++ src/js/core/07-render-charts.js
