+++ src/js/core/10-reset-modal.js
