+++ src/js/core/14-merge-modal.js
