+++ src/js/crypto/11-wallet-popup.js
