+++ src/js/crypto/18b-chain-apply.js
